+++ src/utils/gc/gc_wrappers.c
@@ -32,8 +32,7 @@ char	*gc_strjoin(char const *s1, char const *s2, t_gc_type type)
 	return (res);
 }
 
-char	*gc_substr(char const *s, unsigned int start, size_t len,
-					t_gc_type type)
+char	*gc_substr(char const *s, int start, int len, t_gc_type type)
 {
 	char	*res;
